# internal/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_api.MaintenanceResponse:
    properties:
      affectedGroups:
        items:
          type: string
        type: array
      createdAt:
        type: string
      description:
        type: string
      endTime:
        type: string
      id:
        type: string
      severity:
        type: string
      startTime:
        type: string
      status:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
  description: Self-hosted uptime monitoring API by Project Helena.
  title: Warden API
  version: "1.0"
paths:
...
  /maintenance:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api.MaintenanceResponse'
            type: array
        "500":
          description: Failed to fetch maintenance events
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List maintenance windows
      tags:
      - maintenance
    post:
      consumes:
      - application/json
      parameters:
      - description: Maintenance payload
        in: body
        name: body
        required: true
        schema:
          properties:
            affectedGroups:
              items:
                type: string
              type: array
            description:
              type: string
            endTime:
              type: string
            startTime:
              type: string
            status:
              type: string
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api.MaintenanceResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Failed to schedule maintenance
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create maintenance
      tags:
      - maintenance
  /maintenance/{id}:
    delete:
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              success:
                type: boolean
            type: object
        "400":
          description: Maintenance ID required
          schema:
            type: string
        "500":
          description: Failed to delete maintenance
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete maintenance
      tags:
      - maintenance
    put:
      consumes:
      - application/json
      parameters:
      - description: Maintenance ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated maintenance
        in: body
        name: body
        required: true
        schema:
          properties:
            affectedGroups:
              items:
                type: string
              type: array
            description:
              type: string
            endTime:
              type: string
            startTime:
              type: string
            status:
              type: string
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.MaintenanceResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Failed to update maintenance
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update maintenance
      tags:
      - maintenance
...
  /settings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get settings
      tags:
      - settings
    patch:
      consumes:
      - application/json
      parameters:
      - description: Key-value pairs to update
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              status:
                type: string
            type: object
        "400":
          description: Invalid body
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update settings
      tags:
      - settings
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer sk_live_..." — create keys in Settings > API Keys
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"